Add files via upload
I have created SQL queries for all the tables provided in the project, which include insertion, selection, update, and delete operations. These queries are designed to interact with the death literacy web application database, which contains tables for Responses, Articles, Events, Quizzes, Categories, Comments, Registrations, UserAnswers, and others. I chose to organize the queries into a single .sql file to maintain a clear and streamlined structure.By keeping everything in a single .sql file, we ensure that the queries for interacting with the database are easy to find, execute, and maintain as the project progresses.